fix stateSlice type errors for build, do not export stateSlice in the future
