+++ packages/grafana-prometheus/src/querybuilder/components/metrics-modal/state/helpers.ts
@@ -7,11 +7,12 @@ import { getMetadataHelp, getMetadataType } from '../../../../language_provider'
 import { regexifyLabelValuesQueryString } from '../../../parsingUtils';
 import { QueryBuilderLabelFilter } from '../../../shared/types';
 import { PromVisualQuery } from '../../../types';
+import { setFilteredMetricCount } from '../MetricsModal';
 import { HaystackDictionary, MetricData, MetricsData, PromFilterOption } from '../types';
 
-import { MetricsModalMetadata, MetricsModalState, stateSlice } from './state';
+import { MetricsModalMetadata, MetricsModalState } from './state';
 
-const { setFilteredMetricCount } = stateSlice.actions;
+// const { setFilteredMetricCount } = stateSlice.actions;
 
 export async function setMetrics(
   datasource: PrometheusDatasource,
