+++ packages/grafana-prometheus/src/querybuilder/components/metrics-modal/state/state.ts
@@ -1,5 +1,3 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
 import { SelectableValue } from '@grafana/data';
 
 import { PromVisualQuery } from '../../../types';
@@ -7,81 +5,6 @@ import { HaystackDictionary, MetricsData } from '../types';
 
 export const DEFAULT_RESULTS_PER_PAGE = 100;
 export const MAXIMUM_RESULTS_PER_PAGE = 1000;
-
-export const stateSlice = createSlice({
-  name: 'metrics-modal-state',
-  initialState: initialState(),
-  reducers: {
-    filterMetricsBackend: (
-      state,
-      action: PayloadAction<{
-        metrics: MetricsData;
-        filteredMetricCount: number;
-        isLoading: boolean;
-      }>
-    ) => {
-      state.metrics = action.payload.metrics;
-      state.filteredMetricCount = action.payload.filteredMetricCount;
-      state.isLoading = action.payload.isLoading;
-    },
-    buildMetrics: (state, action: PayloadAction<MetricsModalMetadata>) => {
-      state.isLoading = action.payload.isLoading;
-      state.metrics = action.payload.metrics;
-      state.hasMetadata = action.payload.hasMetadata;
-      state.metaHaystackDictionary = action.payload.metaHaystackDictionary;
-      state.nameHaystackDictionary = action.payload.nameHaystackDictionary;
-      state.totalMetricCount = action.payload.totalMetricCount;
-      state.filteredMetricCount = action.payload.filteredMetricCount;
-    },
-    setIsLoading: (state, action: PayloadAction<boolean>) => {
-      state.isLoading = action.payload;
-    },
-    setFilteredMetricCount: (state, action: PayloadAction<number>) => {
-      state.filteredMetricCount = action.payload;
-    },
-    setResultsPerPage: (state, action: PayloadAction<number>) => {
-      state.resultsPerPage = action.payload;
-    },
-    setPageNum: (state, action: PayloadAction<number>) => {
-      state.pageNum = action.payload;
-    },
-    setFuzzySearchQuery: (state, action: PayloadAction<string>) => {
-      state.fuzzySearchQuery = action.payload;
-      state.pageNum = 1;
-    },
-    setNameHaystack: (state, action: PayloadAction<string[][]>) => {
-      state.nameHaystackOrder = action.payload[0];
-      state.nameHaystackMatches = action.payload[1];
-    },
-    setMetaHaystack: (state, action: PayloadAction<string[][]>) => {
-      state.metaHaystackOrder = action.payload[0];
-      state.metaHaystackMatches = action.payload[1];
-    },
-    setFullMetaSearch: (state, action: PayloadAction<boolean>) => {
-      state.fullMetaSearch = action.payload;
-      state.pageNum = 1;
-    },
-    setIncludeNullMetadata: (state, action: PayloadAction<boolean>) => {
-      state.includeNullMetadata = action.payload;
-      state.pageNum = 1;
-    },
-    setSelectedTypes: (state, action: PayloadAction<Array<SelectableValue<string>>>) => {
-      state.selectedTypes = action.payload;
-      state.pageNum = 1;
-    },
-    setUseBackend: (state, action: PayloadAction<boolean>) => {
-      state.useBackend = action.payload;
-      state.fullMetaSearch = false;
-      state.pageNum = 1;
-    },
-    setDisableTextWrap: (state) => {
-      state.disableTextWrap = !state.disableTextWrap;
-    },
-    showAdditionalSettings: (state) => {
-      state.showAdditionalSettings = !state.showAdditionalSettings;
-    },
-  },
-});
 
 /**
  * Initial state for the metrics explorer
