+++ packages/grafana-prometheus/src/querybuilder/components/promQail/state/helpers.ts
@@ -8,6 +8,7 @@ import { getMetadataHelp, getMetadataType } from '../../../../language_provider'
 import { promQueryModeller } from '../../../PromQueryModeller';
 import { buildVisualQueryFromString } from '../../../parsing';
 import { PromVisualQuery } from '../../../types';
+import { updateInteraction } from '../PromQail';
 import {
   ExplainSystemPrompt,
   GetExplainUserPrompt,
@@ -17,13 +18,11 @@ import {
 } from '../prompts';
 import { Interaction, QuerySuggestion, SuggestionType } from '../types';
 
-import { createInteraction, stateSlice } from './state';
+import { createInteraction } from './state';
 import { getTemplateSuggestions } from './templates';
 
 const OPENAI_MODEL_NAME = 'gpt-3.5-turbo-1106';
 const promQLTemplatesCollection = 'grafana.promql.templates';
-// actions to update the state
-const { updateInteraction } = stateSlice.actions;
 
 interface TemplateSearchResult {
   description: string | null;
